Refactor: rename UI modules for consistent naming
